Selective attribute setting of new documents - set via att writers or raw atts themselves.
This change allows attribute writers to perform secondary functions e.g. validate on write (not yet implemented) or derive other properties. Such behaviour would be undesirable when loading from CouchDB.

I'm not convinced by the use of the set_via_writers constructor param as it exposes an implementation detail to clients that override initialize. Testing for _rev may be a viable alternative.